refactor: KMS 태그 스타일 로직 단순화
