+++ src/components/admin/kms/KmsFeed.vue
@@ -1,7 +1,6 @@
 <script setup>
 import { computed } from 'vue'
 import { BaseFilterTabs } from '@/components/common/base'
-import { getKmsTagStyle } from '@/utils/kmsTagStyle'
 
 const props = defineProps({
   articles:          { type: Array,  default: () => [] },
@@ -10,6 +9,29 @@ const props = defineProps({
   tagFilters:        { type: Array,  default: () => [] },
 })
 const emit = defineEmits(['filterChange', 'tagFilterChange', 'delete', 'restore', 'open-detail', 'toggle-bookmark'])
+
+const TAG_PALETTE = [
+  {
+    bg: 'var(--color-success-soft, #dcfce7)',
+    color: 'var(--color-success-text, #028a6b)',
+  },
+  {
+    bg: 'var(--color-primary-100, #efeaff)',
+    color: 'var(--color-primary-700, #5b4fcf)',
+  },
+  {
+    bg: 'var(--color-warning-soft, #fef3c7)',
+    color: 'var(--color-warning-text, #a07000)',
+  },
+  {
+    bg: 'var(--color-danger-bg, #ffecf1)',
+    color: 'var(--color-danger-text, #c0103e)',
+  },
+  {
+    bg: 'var(--color-bg-surface-muted, #f8f7ff)',
+    color: 'var(--color-text-muted, #7a6fa8)',
+  },
+]
 
 // 카테고리 탭 (백엔드 기준)
 const KMS_FILTERS = [
@@ -36,9 +58,25 @@ const filteredCards = computed(() => {
   return list
 })
 
+function getPaletteIndex(tag) {
+  const value = String(tag ?? '')
+  let hash = 0
+  for (let index = 0; index < value.length; index += 1) {
+    hash = (hash * 31 + value.charCodeAt(index)) % TAG_PALETTE.length
+  }
+  return Math.abs(hash) % TAG_PALETTE.length
+}
+
+function getKmsTagStyle(tag) {
+  if (!tag) {
+    return TAG_PALETTE[0]
+  }
+  return TAG_PALETTE[getPaletteIndex(tag)]
+}
+
 function tagStyle(tag) {
   const found = props.tagFilters.find((t) => t.key === tag)
-  if (found) return { background: found.bg, color: found.color }
+  if (found?.bg && found?.color) return { background: found.bg, color: found.color }
   return getKmsTagStyle(tag)
 }
 </script>
